refactor(fe): move graphql queries to graphql folder
