+++ apps/frontend/app/admin/contest/_components/Columns.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { gql } from '@generated'
 import { DataTableColumnHeader } from '@/components/DataTableColumnHeader'
 import { Checkbox } from '@/components/ui/checkbox'
 import { Switch } from '@/components/ui/switch'
@@ -10,6 +9,7 @@ import {
   TooltipProvider,
   TooltipTrigger
 } from '@/components/ui/tooltip'
+import { EDIT_VISIBLE } from '@/graphql/contest/mutations'
 import { cn, dateFormatter } from '@/lib/utils'
 import { useMutation } from '@apollo/client'
 import * as TooltipPrimitive from '@radix-ui/react-tooltip'
@@ -27,16 +27,6 @@ interface DataTableContest {
   isVisible: boolean
   isRankVisible: boolean
 }
-
-const EDIT_VISIBLE = gql(`
-  mutation UpdateContestVisible($groupId: Int!, $input: UpdateContestInput!) {
-    updateContest(groupId: $groupId, input: $input) {
-      id
-      isVisible
-      isRankVisible
-    }
-  }
-`)
 
 function VisibleCell({ row }: { row: Row<DataTableContest> }) {
   const [updateVisible] = useMutation(EDIT_VISIBLE)
