+++ apps/frontend/app/admin/problem/[id]/edit/page.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import { gql } from '@generated'
 import CheckboxSelect from '@/components/CheckboxSelect'
 import OptionSelect from '@/components/OptionSelect'
 import TagsSelect from '@/components/TagsSelect'
@@ -16,6 +15,9 @@ import {
 import { ScrollArea, ScrollBar } from '@/components/ui/scroll-area'
 import { Switch } from '@/components/ui/switch'
 import { Textarea } from '@/components/ui/textarea'
+import { UPDATE_PROBLEM } from '@/graphql/problem/mutations'
+import { GET_PROBLEM } from '@/graphql/problem/queries'
+import { GET_TAGS } from '@/graphql/problem/queries'
 import { languages, levels } from '@/lib/constants'
 import { cn } from '@/lib/utils'
 import { useMutation, useQuery } from '@apollo/client'
@@ -34,78 +36,7 @@ import { toast } from 'sonner'
 import { z } from 'zod'
 import ExampleTextarea from '../../_components/ExampleTextarea'
 import Label from '../../_components/Label'
-import { GET_TAGS, inputStyle } from '../../utils'
-
-const GET_PROBLEM = gql(`
-  query GetProblem($groupId: Int!, $id: Int!) {
-    getProblem(groupId: $groupId, id: $id) {
-      title
-      isVisible
-      difficulty
-      languages
-      tag {
-        tag {
-          id
-          name
-        }
-      }
-      description
-      inputDescription
-      outputDescription
-      samples {
-        id
-        input
-        output
-      }
-      testcase {
-        id
-        input
-        output
-      }
-      timeLimit
-      memoryLimit
-      hint
-      source
-      template
-    }
-  }
-`)
-
-const UPDATE_PROBLEM = gql(`
-  mutation UpdateProblem($groupId: Int!, $input: UpdateProblemInput!) {
-    updateProblem(groupId: $groupId, input: $input) {
-      id
-      createdById
-      groupId
-      title
-      isVisible
-      difficulty
-      languages
-      problemTag {
-        tag {
-          id
-          name
-        }
-      }
-      description
-      inputDescription
-      outputDescription
-      samples {
-        input
-        output
-      }
-      problemTestcase {
-        input
-        output
-      }
-      timeLimit
-      memoryLimit
-      hint
-      source
-      template
-    }
-  }
-`)
+import { inputStyle } from '../../utils'
 
 const schema = z.object({
   id: z.number(),
