+++ apps/frontend/app/admin/problem/_components/Columns.tsx
@@ -1,8 +1,8 @@
-import { gql } from '@generated'
 import { DataTableColumnHeader } from '@/components/DataTableColumnHeader'
 import { Badge } from '@/components/ui/badge'
 import { Checkbox } from '@/components/ui/checkbox'
 import { Switch } from '@/components/ui/switch'
+import { EDIT_VISIBLE } from '@/graphql/problem/mutations'
 import { useMutation } from '@apollo/client'
 import type { ColumnDef, Row } from '@tanstack/react-table'
 import { FiEyeOff } from 'react-icons/fi'
@@ -24,15 +24,6 @@ interface DataTableProblem {
   languages: string[]
   tag: { id: number; tag: Tag }[]
 }
-
-const EDIT_VISIBLE = gql(`
-  mutation UpdateVisible($groupId: Int!, $input: UpdateProblemInput!) {
-    updateProblem(groupId: $groupId, input: $input) {
-      id
-      isVisible
-    }
-  }
-`)
 
 function VisibleCell({ row }: { row: Row<DataTableProblem> }) {
   const [updateVisible] = useMutation(EDIT_VISIBLE)
